requirements.txt
streamlit
PyPDF2
python-dotenv
requests
phi
openai
pandas
plotly